Checking if the API key is 51 characters long
An easy way to check if the key is a valid API key is to check if it is 51 characters long. If it isn't, it isn't a valid OpenAI key. We still don't really know if the key is valid, but it helps with people typing/copying the wrong amount of characters etc...