Delete OI classic archive
The archive should just be in the git history, not in a folder.  The
classic version can be checked out as a worktree if needed. Having it in
a folder clutters up the repo and is confusing when doing code-wide
searches, finding seemingly-relevant references that are actually
obsolete, or when using tools like Cursor that look for things
throughout the repo and get confused about new and old things with the
same name.